micro/requirements.txt
Django==5.2.10
pymysql==1.1.0
djangorestframework==3.16.1
django-cors-headers==4.6.0
requests==2.32.3
